rm_pm_hal: fix PRECISION.LOSS.CALL
Instances in which an implicit cast to a smaller data type during
a function call can cause a loss of precision in data.

Change of size from u8 to u16.

Fixes: SYSFW-6265

+++ pm/drivers/clock/clk_mux.c
@@ -179,7 +179,7 @@ const struct clk_parent *clk_get_parent(struct clk *clkp)
 }
 
 /* FIXME: freq change ok/notify? new freq in range? */
-sbool clk_set_parent(struct clk *clkp, u8 new_parent)
+sbool clk_set_parent(struct clk *clkp, u16 new_parent)
 {
 	const struct clk_data *clk_datap = clk_get_data(clkp);
 	const struct clk_drv_mux *mux_drv = NULL;
